fix: working_times 널체크 개선

+++ backend/store/serializers.py
@@ -349,16 +349,20 @@ class OwnerStoreUpdateSerializer(serializers.ModelSerializer):
 
         # 운영시간 전달 시 전체 교체(upsert)
         if working_times is not None:
-            instance.storeworkingtime_set.all().delete()
-            working_times = [
-                StoreWorkingTime(
-                    store_id=instance,
-                    working_day=f'D{i:02d}',
-                    start_time=working_times.get('start_time'),
-                    end_time=working_times.get('end_time')
-                ) for i in range(1, 8)  # 월(D01) ~ 일(D07)
-            ]
-            StoreWorkingTime.objects.bulk_create(working_times)
+            start_time = working_times.get('start_time')  # ← 먼저 값 추출
+            end_time = working_times.get('end_time')
+
+            if start_time is not None and end_time is not None:
+                instance.storeworkingtime_set.all().delete()
+                working_days = [
+                    StoreWorkingTime(
+                        store_id=instance,
+                        working_day=f'D{i:02d}',
+                        start_time=start_time,
+                        end_time=end_time
+                    ) for i in range(1, 8)  # 월(D01) ~ 일(D07)
+                ]
+                StoreWorkingTime.objects.bulk_create(working_days)
 
         if off_dates is not None :
             for od in off_dates:
